a few minor typos remedied
I didn't want to change this because I've not made much use of the every function, but this paragraph:

"In Scheme, we usually have to be a bit more specific. It is customary
to define two functions, usually called \texttt{every} and \texttt{any},
that take a predicate and a list of objects from the domain, and evaluate
to \texttt{\#false} if every object from the list satisfies the
predicate (in case of \texttt{every}) or there is no object in the list
that satisfies the predicate (in case of \texttt{and}), and otherwise
evaluate to some truth-ish value."

shouldn't that be #true if every object from the list satisfies the predicate?
also, shouldn't "(in the case of and)" be "(in the case of any)" ?

interesting read. thanks again for linking me here from on the newgroup :)